Реализует отрисовку заголовка при получении списка фильмов
